requirements.txt
reportlab==4.0.4
svglib==1.5.1
Pillow==10.0.0
PyPDF2==3.0.1
pytest==7.4.0
pytest-cov==4.1.0 